+++ recipes/default.rb
@@ -145,7 +145,7 @@ end
 # Deploy additional scripts, preferably outside Android-SDK own directories to
 # avoid unwanted removal when updating android sdk components later.
 #
-%w(android-accept-licenses).each do |android_helper_script|
+%w(android-accept-licenses android-wait-for-emulator).each do |android_helper_script|
   cookbook_file File.join(node['android-sdk']['scripts']['path'], android_helper_script) do
     source android_helper_script
     owner  node['android-sdk']['scripts']['owner']