+++ CHANGELOG.md
@@ -6,6 +6,7 @@ v0.2.0 (upcoming)
 
 - Integrate by default with Android SDK 23.0.0
 - Optionally install Maven Android SDK Deployer [GH-14]
+- Deploy scripts for waiting on Emulator startup [GH-16]
 - Deploy scripts for non-interactive SDK setup/updates [GH-13]
 
 v0.1.1 (2014-04-01)